Fix bug of fetching traces
