+++ zipkin-lens/src/components/GlobalSearch/util.js
@@ -77,75 +77,3 @@ export const retrieveDefaultConditionValue = (conditionKey) => {
       return undefined;
   }
 };
-
-export const extractConditionsFromQueryParameters = (queryParameters) => {
-  const conditions = [];
-  const lookbackCondition = {};
-  let limitCondition = 0;
-
-  Object.keys(queryParameters).forEach((conditionKey) => {
-    const conditionValue = queryParameters[conditionKey];
-    switch (conditionKey) {
-      case 'serviceName':
-      case 'remoteServiceName':
-      case 'spanName':
-        conditions.push({
-          key: conditionKey,
-          value: conditionValue,
-        });
-        break;
-      case 'minDuration':
-      case 'maxDuration':
-        conditions.push({
-          key: conditionKey,
-          value: parseInt(conditionValue, 10),
-        });
-        break;
-      case 'tags':
-        conditionValue.split(' and ').forEach((tags) => {
-          conditions.push({
-            key: 'tags',
-            value: tags,
-          });
-        });
-        break;
-      case 'autocompleteTags':
-        conditionValue.split(' and ').forEach((autocompleteTag) => {
-          const splitted = autocompleteTag.split('=');
-          conditions.push({
-            key: splitted[0],
-            value: splitted[1],
-          });
-        });
-        break;
-      case 'limit':
-        limitCondition = parseInt(conditionValue, 10);
-        break;
-      case 'lookback':
-        switch (conditionValue) {
-          case '1h':
-          case '2h':
-          case '6h':
-          case '12h':
-          case '1d':
-          case '2d':
-          case '7d': {
-            lookbackCondition.value = conditionValue;
-            lookbackCondition.endTs = parseInt(queryParameters.endTs, 10);
-            break;
-          }
-          case 'custom':
-            lookbackCondition.value = conditionValue;
-            lookbackCondition.endTs = parseInt(queryParameters.endTs, 10);
-            lookbackCondition.startTs = parseInt(queryParameters.startTs, 10);
-            break;
-          default:
-            break;
-        }
-        break;
-      default:
-        break;
-    }
-  });
-  return { conditions, lookbackCondition, limitCondition };
-};
